+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Router from './router';
+import Router from './router/Router';
 import AuthModal from './components/auth/AuthModal';
 
 function App() {
