--- src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './shared';
-export * from './resume/ResumeForm';
