--- src/components/shared/Button/types.ts
@@ -1,10 +1,0 @@
-export type ButtonVariant = 'primary' | 'secondary' | 'outline' | 'white';
-
-export interface ButtonProps {
-  children: React.ReactNode;
-  type?: 'button' | 'submit';
-  variant?: ButtonVariant;
-  onClick?: () => void;
-  disabled?: boolean;
-  className?: string;
-}
