--- src/components/shared/Dropdown/types.ts
@@ -1,12 +1,0 @@
-export interface DropdownProps<T = string> {
-  label?: string;
-  name?: string;
-  items: DropdownItem<T>[];
-  selected?: T;
-  onSelect: (value: T) => void;
-}
-
-export interface DropdownItem<T = string> {
-  name: string;
-  value: T;
-}
