+++ src/components/shared/InputField/index.tsx
@@ -1,5 +1,16 @@
-import { InputFieldProps } from './types';
 import clsx from 'clsx';
+
+interface InputFieldProps {
+  type?: string;
+  label?: string;
+  name: string;
+  value?: string;
+  placeholder?: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  error?: string;
+  disabled?: boolean;
+  className?: string;
+}
 
 export default function InputField({
   type,
