--- src/components/shared/InputField/types.ts
@@ -1,11 +1,0 @@
-export interface InputFieldProps {
-  type?: string;
-  label?: string;
-  name: string;
-  value?: string;
-  placeholder?: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  error?: string;
-  disabled?: boolean;
-  className?: string;
-}
