--- src/components/shared/MultiDropdown/types.ts
@@ -1,10 +1,0 @@
-import { DropdownItem } from '../Dropdown/types';
-
-export interface MultiDropdownProps<T = string> {
-  label?: string;
-  name?: string;
-  items: DropdownItem<T>[];
-  selcted?: T;
-  value?: string;
-  onChange: (val: string) => void;
-}
