--- src/components/shared/MultiSelectInput/types.ts
@@ -1,8 +1,0 @@
-export interface MultiSelectInputProps {
-  label?: string;
-  selected: string[];
-  onChange: (items: string[]) => void;
-  placeholder?: string;
-  error?: string;
-  className?: string;
-}
