--- src/components/shared/TextArea/types.ts
@@ -1,8 +1,0 @@
-export interface TextAreaProps {
-  // label?: string;
-  value: string;
-  onChange: (value: string) => void;
-  placeholder?: string;
-  maxLength?: number;
-  className?: string;
-}
