+++ src/components/shared/index.ts
@@ -1,6 +1,8 @@
+// @/components/shared 로 한 번에 호출 가능
 export { default as Button } from './Button';
 export { default as Dropdown } from './Dropdown';
 export { default as InputField } from './InputField';
+export { default as Loading } from './Loading';
 export { default as MultiDropdown } from './MultiDropdown';
 export { default as MultiSelectInput } from './MultiSelectInput';
 export { default as TextArea } from './TextArea';
