+++ src/pages/guest/Guest.tsx
@@ -1,5 +1,5 @@
 import { postGuestCountry } from '@/api/guest';
-import { Button } from '@/components';
+import { Button } from '@/components/shared';
 import Modal from '@/components/layout/Modal';
 import ResumeForm from '@/components/resume/ResumeForm';
 import Loading from '@/components/shared/Loading';
